pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["mcp_github"]

[project]
name = "mcp-github-fastmcp"
version = "0.1.0"
description = "GitHub MCP server built with fastMCP"
readme = "README.md"
requires-python = ">=3.11"

authors = [
    {name = "Hyunwoo Jang", email = "wkdgusdn0321@naver.com"}
]
keywords = ["mcp", "github", "fastmcp"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "fastmcp",
    "python-dotenv",
    "rich",
    "pydantic",
    "PyGithub",
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "httpx>=0.25.2",
    "tenacity>=8.2.3",
    "structlog>=23.2.0",
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
    "mypy",
    "pre-commit",
    "commitizen",
    "pytest-cov",
    "black",
    "isort",
    "pyupgrade",
    "bandit",
    "safety",
    "pdoc",
    "pytest-asyncio>=0.21.1",
]

[project.scripts]
mcp-github = "mcp_github.server:main"

[tool.ruff]
target-version = "py311"
line-length = 88
select = ["E", "F", "I", "N", "W", "B", "C4", "UP"]
ignore = ["E501", "B008"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v"
asyncio_mode = "auto"

[tool.commitizen]
name = "cz_conventional_commits"
version = "0.1.0"
tag_format = "v$version"
